requirements.txt
fastapi
"uvicorn[standard]"
python-dotenv
httpx
jinja2
